\name{ajusta_modelo_fatorial}
\alias{ajusta_modelo_fatorial}
\title{ajusta_modelo_fatorial}
\usage{
ajusta_modelo_fatorial(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

