\name{anova_diagnostico}
\alias{anova_diagnostico}
\title{anova_diagnostico}
\usage{
anova_diagnostico(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

