\name{colunas_com_na}
\alias{colunas_com_na}
\title{colunas_com_na}
\usage{
colunas_com_na(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

