\name{configurar_ambiente_rlib}
\alias{configurar_ambiente_rlib}
\title{configurar_ambiente_rlib}
\usage{
configurar_ambiente_rlib(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

