% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_funcoes_graficos.R
\name{grafico_interacao_fatorial}
\alias{grafico_interacao_fatorial}
\title{Grafico de interacao para dois fatores}
\usage{
grafico_interacao_fatorial(
  dados,
  resposta,
  fator_x,
  fator_traco,
  bloco = NULL,
  fatores,
  dic_vars = NULL,
  label_type = c("label", "sigla", "description", "var"),
  digitos = 2
)
}
\arguments{
\item{dados}{\code{data.frame} com os dados experimentais.}

\item{resposta}{Nome da variavel resposta.}

\item{fator_x}{Fator no eixo x.}

\item{fator_traco}{Fator representado por cor/linha.}

\item{bloco}{Nome do bloco (DBC). Use \code{NULL} para DIC.}

\item{fatores}{Vetor de nomes dos fatores do modelo.}

\item{dic_vars}{Dicionario opcional com colunas \code{var}, \code{sigla}, \code{label}, \code{description}.}

\item{label_type}{Tipo de rotulo da variavel resposta.}

\item{digitos}{Numero de casas decimais.}
}
\value{
Objeto \code{ggplot2}.
}
\description{
Grafico de interacao para dois fatores
}
