\name{grafico_media_ep}
\alias{grafico_media_ep}
\title{grafico_media_ep}
\usage{
grafico_media_ep(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

