\name{grafico_multiplas_variaveis}
\alias{grafico_multiplas_variaveis}
\title{grafico_multiplas_variaveis}
\usage{
grafico_multiplas_variaveis(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

