% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_funcoes_utilitarias.R
\name{manter_colunas_fixas_e_com_na}
\alias{manter_colunas_fixas_e_com_na}
\title{Mantem colunas fixas e colunas com valores ausentes}
\usage{
manter_colunas_fixas_e_com_na(dados, colunas_fixas)
}
\arguments{
\item{dados}{\code{data.frame} de entrada.}

\item{colunas_fixas}{Vetor de nomes de colunas que devem ser mantidas.}
}
\value{
\code{data.frame} filtrado pelas colunas solicitadas.
}
\description{
Mantem colunas fixas e colunas com valores ausentes
}
