% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranova-package.R
\docType{package}
\name{ranova-package}
\alias{ranova}
\alias{ranova-package}
\title{ranova: Analise de Variancia para Experimentos}
\description{
Ferramentas para ajustar modelos fatoriais (DIC e DBC), gerar ANOVA,
tabelas de medias, diagnostico de pressupostos e graficos para relatorios.
}
\author{
\strong{Maintainer}: Marlenildo Ferreira Melo \email{marlenildo@example.com}

}
\keyword{internal}
