% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_funcoes_utilitarias.R
\name{remove_colunas_com_na}
\alias{remove_colunas_com_na}
\title{Remove colunas com qualquer valor ausente}
\usage{
remove_colunas_com_na(dados)
}
\arguments{
\item{dados}{\code{data.frame} de entrada.}
}
\value{
\code{data.frame} sem colunas que possuam \code{NA}.
}
\description{
Remove colunas com qualquer valor ausente
}
