% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_funcoes_utilitarias.R
\name{remove_colunas_todas_na}
\alias{remove_colunas_todas_na}
\title{Remove colunas totalmente ausentes}
\usage{
remove_colunas_todas_na(dados)
}
\arguments{
\item{dados}{\code{data.frame} de entrada.}
}
\value{
\code{data.frame} sem colunas onde todos os valores sao \code{NA}.
}
\description{
Remove colunas totalmente ausentes
}
