\name{resolve_var_label}
\alias{resolve_var_label}
\title{resolve_var_label}
\usage{
resolve_var_label(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

