% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlibfatorial-package.R
\docType{package}
\name{rlibfatorial-package}
\alias{rlibfatorial}
\alias{rlibfatorial-package}
\title{rlibfatorial: Analise de Experimentos Fatoriais}
\description{
Ferramentas para ajustar modelos fatoriais (DIC e DBC), gerar ANOVA,
tabelas de medias, diagnostico de pressupostos e graficos para relatorios.
}
\author{
\strong{Maintainer}: Marlenildo Ferreira Melo \email{marlenildo@example.com}

}
\keyword{internal}
