\name{seleciona_colunas_com_na}
\alias{seleciona_colunas_com_na}
\title{seleciona_colunas_com_na}
\usage{
seleciona_colunas_com_na(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

