\name{tabela_dt_exportavel}
\alias{tabela_dt_exportavel}
\title{tabela_dt_exportavel}
\usage{
tabela_dt_exportavel(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

