% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_funcoes_modelo_fatorial_medias.R
\name{ajusta_modelo_fatorial}
\alias{ajusta_modelo_fatorial}
\title{Ajusta modelo fatorial (DIC ou DBC)}
\usage{
ajusta_modelo_fatorial(dados, resposta, bloco = NULL, fatores)
}
\arguments{
\item{dados}{\code{data.frame} com os dados experimentais.}

\item{resposta}{Nome da variavel resposta.}

\item{bloco}{Nome da coluna de blocos (DBC). Use \code{NULL} para DIC.}

\item{fatores}{Vetor com nomes dos fatores.}
}
\value{
Objeto \code{aov}.
}
\description{
Ajusta modelo fatorial (DIC ou DBC)
}
