\name{anova_fatorial_qm_tabela}
\alias{anova_fatorial_qm_tabela}
\title{anova_fatorial_qm_tabela}
\usage{
anova_fatorial_qm_tabela(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

