% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_funcoes_utilitarias.R
\name{colunas_com_na}
\alias{colunas_com_na}
\title{Lista nomes de colunas com valores ausentes}
\usage{
colunas_com_na(dados)
}
\arguments{
\item{dados}{\code{data.frame} de entrada.}
}
\value{
Vetor de caracteres com nomes de colunas contendo \code{NA}.
}
\description{
Lista nomes de colunas com valores ausentes
}
