\name{grafico_interacao_fatorial}
\alias{grafico_interacao_fatorial}
\title{grafico_interacao_fatorial}
\usage{
grafico_interacao_fatorial(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

