\name{manter_colunas_fixas_e_com_na}
\alias{manter_colunas_fixas_e_com_na}
\title{manter_colunas_fixas_e_com_na}
\usage{
manter_colunas_fixas_e_com_na(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

