\name{medias_fatorial_cld}
\alias{medias_fatorial_cld}
\title{medias_fatorial_cld}
\usage{
medias_fatorial_cld(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

