\name{remove_colunas_todas_na}
\alias{remove_colunas_todas_na}
\title{remove_colunas_todas_na}
\usage{
remove_colunas_todas_na(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

