% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_funcoes_modelo_fatorial_medias.R
\name{resolve_var_label}
\alias{resolve_var_label}
\title{Resolve rotulo de variavel por dicionario}
\usage{
resolve_var_label(var, dic_vars = NULL, type = c("label", "sigla", "var"))
}
\arguments{
\item{var}{Nome da variavel.}

\item{dic_vars}{Dicionario opcional com colunas \code{var}, \code{label}, \code{sigla}.}

\item{type}{Tipo de rotulo retornado.}
}
\value{
Vetor de caracteres com o rotulo resolvido.
}
\description{
Resolve rotulo de variavel por dicionario
}
