% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_funcoes_utilitarias.R
\name{seleciona_colunas_com_na}
\alias{seleciona_colunas_com_na}
\title{Seleciona apenas colunas com algum valor ausente}
\usage{
seleciona_colunas_com_na(dados)
}
\arguments{
\item{dados}{\code{data.frame} de entrada.}
}
\value{
\code{data.frame} contendo somente colunas com \code{NA}.
}
\description{
Seleciona apenas colunas com algum valor ausente
}
