% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_funcoes_tabelas.R
\name{tabela_dt_exportavel}
\alias{tabela_dt_exportavel}
\title{Cria tabela interativa exportavel via DataTables}
\usage{
tabela_dt_exportavel(dados, titulo = "Visualizacao dos dados")
}
\arguments{
\item{dados}{\code{data.frame} a ser exibido.}

\item{titulo}{Titulo mostrado na tabela.}
}
\value{
Objeto \code{DT::datatable}.
}
\description{
Cria tabela interativa exportavel via DataTables
}
