\name{tabela_interacao_fatorial}
\alias{tabela_interacao_fatorial}
\title{tabela_interacao_fatorial}
\usage{
tabela_interacao_fatorial(...)
}
\arguments{
  \item{...}{Consulte o codigo-fonte da funcao para os argumentos.}
}
\value{Objeto retornado pela funcao.}
\description{Funcao exportada do pacote rlibfatorial para analise fatorial e relatorios.}

