% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_funcoes_modelo_fatorial_medias.R
\name{tabela_medias_fatorial}
\alias{tabela_medias_fatorial}
\title{Monta tabela de medias fatoriais}
\usage{
tabela_medias_fatorial(
  dados,
  variaveis,
  fator_interesse,
  bloco = NULL,
  fatores,
  dic_vars = NULL,
  label_type = c("label", "sigla", "var"),
  caption = "MACdias ajustadas A? erro-padrA?o.",
  digitos = 2,
  tipo_se = c("modelo", "descritivo")
)
}
\arguments{
\item{dados}{\code{data.frame} com os dados experimentais.}

\item{variaveis}{Vetor de nomes das variaveis resposta.}

\item{fator_interesse}{Nome do fator para comparacao.}

\item{bloco}{Nome da coluna de blocos (DBC). Use \code{NULL} para DIC.}

\item{fatores}{Vetor com nomes dos fatores.}

\item{dic_vars}{Dicionario opcional com colunas \code{var}, \code{label}, \code{sigla}.}

\item{label_type}{Tipo de rotulo para variaveis resposta.}

\item{caption}{Legenda da tabela.}

\item{digitos}{Numero de casas decimais.}

\item{tipo_se}{Tipo de erro-padrao (\code{modelo} ou \code{descritivo}).}
}
\value{
Objeto \code{knitr_kable}.
}
\description{
Monta tabela de medias fatoriais
}
